refactor: Move InfoButtons into ui folder
